% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coerce.R
\name{as.data.frame}
\alias{as.data.frame}
\alias{as.data.frame.ModelFrame}
\alias{as.data.frame.Resample}
\alias{as.data.frame.TabularArray}
\title{Coerce to a Data Frame}
\usage{
\method{as.data.frame}{ModelFrame}(x, ...)

\method{as.data.frame}{Resample}(x, ...)

\method{as.data.frame}{TabularArray}(x, ...)
}
\arguments{
\item{x}{\code{\link{ModelFrame}}, \link{resample} results, resampled
\link{performance} estimates, model performance \link[=diff]{differences},
or \link[=t.test]{t-test} comparisons of the differences.}

\item{...}{arguments passed to other methods.}
}
\value{
\code{data.frame} class object.
}
\description{
Functions to coerce objects to data frames.
}
