% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coerce.R
\name{as.MLInput}
\alias{as.MLInput}
\alias{as.MLInput.MLModelFit}
\alias{as.MLInput.ModelSpecification}
\title{Coerce to an MLInput}
\usage{
as.MLInput(x, ...)

\method{as.MLInput}{MLModelFit}(x, ...)

\method{as.MLInput}{ModelSpecification}(x, ...)
}
\arguments{
\item{x}{model \link{fit} result or \pkg{MachineShop}
\link[=ModelSpecification]{model specification}.}

\item{...}{arguments passed to other methods.}
}
\value{
\code{MLInput} class object.
}
\description{
Function to coerce an object to \code{MLInput}.
}
