% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c.R
\name{c}
\alias{c}
\alias{combine}
\alias{c.Resamples}
\title{Combine MachineShop Objects}
\usage{
\method{c}{Resamples}(...)
}
\arguments{
\item{...}{named or unnamed \link{resample} results.  Resamples must have
been generated with the same resampling \link[=controls]{control}.}
}
\value{
Object of the same class as the arguments.
}
\description{
Combine one or more \pkg{MachineShop} objects of the same class.
}
