% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLC50Model.R
\name{C50Model}
\alias{C50Model}
\title{C5.0 Decision Trees and Rule-Based Model}
\usage{
C50Model(trials = NULL, rules = NULL, control = NULL)
}
\arguments{
\item{trials}{integer number of boosting iterations. A value of one indicates
that a single model is used.}

\item{rules}{logical indicating whether the tree is decomposed into a
rule-based model}

\item{control}{list of control parameters.}
}
\value{
MLModel class object.
}
\description{
Fit classification tree models or rule-based models using Quinlan's C5.0
algorithm.
}
\details{
\describe{
\item{Response Types:}{\code{factor}}
}

Default values for the \code{NULL} arguments and further model
details can be found in the source link below.
}
\seealso{
\code{\link[C50]{C5.0}}, \code{\link{fit}}, \code{\link{resample}},
\code{\link{tune}}
}
