% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLCForestModel.R
\name{CForestModel}
\alias{CForestModel}
\title{Conditional Random Forest Model}
\usage{
CForestModel(controls = NULL)
}
\arguments{
\item{controls}{an object of class \code{\linkS4class{ForestControl}}.}
}
\description{
An implementation of the random forest and bagging ensemble algorithms
utilizing conditional inference trees as base learners.
}
\details{
\describe{
\item{Response Types:}{\code{factor}, \code{numeric}, \code{Surv}}
}

Default values for the \code{NULL} arguments and further model
details can be found in the source link below.
}
\seealso{
\code{\link[party]{cforest}}
}
