% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confusion.R
\name{confusion}
\alias{confusion}
\alias{confusion.Resamples}
\alias{confusion.factor}
\title{Confusion Matrix}
\usage{
confusion(x, ...)

\method{confusion}{Resamples}(x, ...)

\method{confusion}{factor}(x, y, cutoff = 0.5, ...)
}
\arguments{
\item{x}{observed responses or class containing observed and predicted
responses.}

\item{...}{arguments passed to other methods.}

\item{y}{predicted responses.}

\item{cutoff}{threshold above which probabilities are classified as success
for binary responses.}
}
\value{
\code{ConfusionResamples} class object that inherits from
\code{list}.

\code{ConfusionMatrix} class object that inherits from \code{table}.
}
\description{
Calculate confusion matrices of predicted and observed responses.
}
\details{
\code{Resamples} responses are summed directly over predicted class
probabilities for the cross-classification and will thus appear as decimal
numbers that can be interpreted as expected counts.
}
\examples{
res <- resample(Species ~ ., data = iris, model = GBMModel)
confusion(res)

}
\seealso{
\code{\link{resample}}, \code{\link{plot}}, \code{\link{summary}}
}
