% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confusion.R
\name{confusion}
\alias{confusion}
\title{Confusion Matrix}
\usage{
confusion(x, ...)
}
\arguments{
\item{x}{Resamples object.}

\item{...}{arguments passed to other methods.}
}
\value{
\code{by} class object that inherits from \code{list}.
}
\description{
Calculate confusion matrices summarizing the proportions of cross-classified
observed and predicted responses
}
\examples{
perf <- resample(Species ~ ., data = iris, model = GBMModel)
confusion(perf)

}
\seealso{
\code{\link{resample}}
}
