% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response.R
\name{DiscreteVector}
\alias{DiscreteVector}
\alias{BinomialMatrix}
\alias{NegBinomialVector}
\alias{PoissonVector}
\title{Discrete Variable Constructors}
\usage{
BinomialMatrix(count = integer(), size = integer())

DiscreteVector(x = integer(), min = -Inf, max = Inf)

NegBinomialVector(x = integer())

PoissonVector(x = integer())
}
\arguments{
\item{count, size}{numeric vectors of binomial counts and trials.}

\item{x}{numeric vector.}

\item{min, max}{minimum and maximum bounds for discrete numbers.}
}
\value{
\code{BinomialMatrix} object class that inherits from \code{matrix},
\code{DiscreteVector} that inherits from \code{numeric}, or object of the
same class as the constructor name and that inherits from
\code{DiscreteVector}.
}
\description{
Create an integer matrix of binomial counts or vector of discrete numbers,
negative binomial counts, or Poisson counts.
}
\examples{
BinomialMatrix(rbinom(25, 10, 0.5), size = 10)
PoissonVector(rpois(25, 10))

}
\seealso{
\code{\link{role_binom}}
}
