% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.}
\alias{.}
\title{Quote Operator}
\usage{
.(expr)
}
\arguments{
\item{expr}{any syntactically valid \R expression.}
}
\value{
The quoted (unevaluated) expression.
}
\description{
Shorthand notation for the \code{\link{quote}} function.  The quote operator
simply returns its argument unevaluated and can be applied to any \R
expression.  Useful for calling model constructors with quoted parameter
values that are defined in terms of a model \code{formula}, \code{data},
\code{weights}, \code{nobs}, \code{nvars}, or \code{y}.
}
\seealso{
\code{\link{quote}}
}
