% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ML_EarthModel.R
\name{EarthModel}
\alias{EarthModel}
\title{Multivariate Adaptive Regression Splines Model}
\usage{
EarthModel(pmethod = c("backward", "none", "exhaustive", "forward",
  "seqrep", "cv"), trace = 0, degree = 1, nprune = NULL, nfold = 0,
  ncross = 1, stratify = TRUE)
}
\arguments{
\item{pmethod}{pruning method.}

\item{trace}{level of execution information to display.}

\item{degree}{maximum degree of interaction.}

\item{nprune}{maximum number of terms (including intercept) in the pruned
model.}

\item{nfold}{number of cross-validation folds.}

\item{ncross}{number of cross-validations if \code{nfold > 1}.}

\item{stratify}{logical indicating whether to stratify cross-validation
samples by the response levels.}
}
\value{
\code{MLModel} class object.
}
\description{
Build a regression model using the techniques in Friedman's papers
"Multivariate Adaptive Regression Splines" and "Fast MARS".
}
\details{
\describe{
\item{Response Types:}{\code{factor}, \code{numeric}}
}

Default values for the \code{NULL} arguments and further model details can be
found in the source link below.

In calls to \code{\link{varimp}} for \code{EarthModel}, argument
\code{metric} may be specified as \code{"gcv"} (default) for the generalized
cross-validation decrease over all subsets that include each predictor, as
\code{"rss"} for the residual sums of squares decrease, or as
\code{"nsubsets"} for the number of model subsets that include each
predictor.  Variable importance is automatically scaled to range from 0 to
100.  To obtain unscaled importance values, set \code{scale = FALSE}.  See
example below.
}
\examples{
modelfit <- fit(Species ~ ., data = iris, model = EarthModel())
varimp(modelfit, metric = "nsubsets", scale = FALSE)

}
\seealso{
\code{\link[earth]{earth}}, \code{\link{fit}},
\code{\link{resample}}, \code{\link{tune}}
}
