% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand.model.R
\name{expand.model}
\alias{expand.model}
\title{Model Expansion Over a Grid of Tuning Parameters}
\usage{
expand.model(x, ...)
}
\arguments{
\item{x}{\link[=models]{model} function, function name, or call.}

\item{...}{vectors, factors, or list containing the parameter values over
which to expand \code{x}.}
}
\value{
A list of model objects created from the parameter combinations.
}
\description{
Expand a model over all combinations of a grid of tuning parameters.
}
\examples{
expand.model(GBMModel, n.trees = c(25, 50, 100),
                       interaction.depth = 1:3,
                       n.minobsinnode = c(5, 10))

}
\seealso{
\code{\link{tune}}
}
