% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_model.R
\name{expand_model}
\alias{expand_model}
\alias{expand.model}
\title{Model Expansion Over a Grid of Tuning Parameters}
\usage{
expand_model(x, ...)

expand.model(...)
}
\arguments{
\item{x}{\code{MLModel} function, function name, or object.}

\item{...}{vectors, factors, or a list containing the parameter values.}
}
\value{
A list of \code{MLModel} objects created from the parameter
combinations.
}
\description{
Expand a model over all combinations of a grid of tuning parameters.
}
\examples{
expand_model(GBMModel, n.trees = c(25, 50, 100),
                       interaction.depth = 1:3,
                       n.minobsinnode = c(5, 10))

}
\seealso{
\code{\link{modelinfo}}, \code{\link{tune}}
}
