% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\docType{methods}
\name{extract}
\alias{extract}
\alias{[,Resamples,ANY,ANY,ANY-method}
\alias{[.SurvMatrix}
\title{Extract Parts of an Object}
\usage{
\S4method{[}{Resamples,ANY,ANY,ANY}(x, i, j, drop = FALSE)

\method{[}{SurvMatrix}(x, i, j, drop = FALSE)
}
\arguments{
\item{x}{object from which to extract elements.}

\item{i, j}{indices specifying elements to extract.}

\item{drop}{logical indicating that the result be returned as a
\code{numeric} coerced to the lowest dimension possible if \code{TRUE} or
retained as the original 2-dimensional object class otherwise.}
}
\description{
Operators acting on data structures to extract parts.
}
\seealso{
\code{\link{resample}}

\code{\link{SurvMatrix}}
}
