% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLGBMModel.R
\name{GBMModel}
\alias{GBMModel}
\title{Generalized Boosted Regression Model}
\usage{
GBMModel(distribution = NULL, n.trees = NULL,
  interaction.depth = NULL, n.minobsinnode = NULL, shrinkage = NULL,
  bag.fraction = NULL)
}
\arguments{
\item{distribution}{either a character string specifying the name of the
distribution to use or a list with a component \code{name} specifying the
distribution and any additional parameters needed.}

\item{n.trees}{total number of trees to fit.}

\item{interaction.depth}{maximum depth of variable interactions.}

\item{n.minobsinnode}{minimum number of observations in the trees terminal
nodes.}

\item{shrinkage}{shrinkage parameter applied to each tree in the expansion.}

\item{bag.fraction}{fraction of the training set observations randomly
selected to propose the next tree in the expansion.}
}
\description{
Fits generalized boosted regression models.
}
\details{
\describe{
\item{Response Types:}{\code{factor}, \code{numeric}, \code{Surv}}
}

Default values for the \code{NULL} arguments and further model
details can be found in the source link below.
}
\seealso{
\code{\link[gbm]{gbm}}
}
