% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{Grid}
\alias{Grid}
\title{Tuning Grid Control}
\usage{
Grid(size = 3, random = FALSE, length = NULL)
}
\arguments{
\item{size}{number of values to be generated for each model parameter in
the tuning grid.}

\item{random}{number of unique grid points to sample at random, \code{Inf}
for all random points, or \code{FALSE} for all fixed points.}

\item{length}{deprecated argument; use \code{size} instead.}
}
\value{
\code{Grid} class object.
}
\description{
Defines control parameters for a tuning grid.
}
\examples{
TunedModel(GBMModel, grid = Grid(10, random = 5))

}
\seealso{
\code{\link{TunedModel}}
}
