% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ICHomes}
\alias{ICHomes}
\title{Iowa City Home Sales Dataset}
\format{A data frame with 753 observations of 17 variables:
\describe{
  \item{sale_amount}{sale amount in dollars.}
  \item{sale_year}{sale year.}
  \item{sale_month}{sale month.}
  \item{built}{year in which the home was built.}
  \item{style}{home stlye (Home/Condo)}
  \item{construction}{home construction type.}
  \item{base_size}{base foundation size in sq ft.}
  \item{add_size}{size of additions made to the base foundation in sq ft.}
  \item{garage1_size}{attached garage size in sq ft.}
  \item{garage2_size}{detached garage size in sq ft.}
  \item{lot_size}{total lot size in sq ft.}
  \item{bedrooms}{number of bedrooms.}
  \item{basement}{presence of a basement (No/Yes).}
  \item{ac}{presence of central air conditioning (No/Yes).}
  \item{attic}{presence of a finished attic (No/Yes).}
  \item{lon,lat}{home longitude/latitude coordinates.}
}}
\usage{
ICHomes
}
\description{
Characteristics of homes sold in Iowa City, IA from 2005 to 2008 as reported
by the county assessor's office.
}
\keyword{datasets}
