% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{MLModel}
\alias{MLModel}
\title{MLModel Class Constructor}
\usage{
MLModel(name = "MLModel", label = name, packages = character(0),
  types = character(0), params = list(), nvars = function(data) NULL,
  fit = function(formula, data, weights, ...) stop("no fit function"),
  predict = function(object, newdata, times, ...)
  stop("no predict function"), varimp = function(object, ...) NULL)
}
\arguments{
\item{name}{character string name for the instantiated \code{MLModel} object;
same as the derived model constructor name.}

\item{label}{descriptive label for the model.}

\item{packages}{character vector of packages required by the object.}

\item{types}{character vector of response variable types on which the model
can be fit.}

\item{params}{list of user-specified model parameters.}

\item{nvars}{function to return the number of predictor variables for a
given model frame.}

\item{fit}{model fitting function.}

\item{predict}{model prediction function.}

\item{varimp}{variable importance function.}
}
\value{
\code{MLModel} class object.
}
\description{
Create a model for use with the \pkg{MachineShop} package.
}
\details{
\code{MLModel} can be used to create custom model constructors.  To
ensure full compatibility with the \pkg{MachineShop} package, add derived
constructor functions to the package namespace with the \code{environment}
function.  See example below.
}
\examples{
## Logistic regression model constructor
LogisticModel <- function() {
  MLModel(
    name = "LogisticModel",
    label = "Logistic Regression",
    package = "stats",
    types = "binary",
    fit = function(formula, data, weights, ...) {
      environment(formula) <- environment()
      glm(formula, data = data, weights = weights, family = binomial, ...)
    },
    predict = function(object, newdata, ...) {
      predict(object, newdata = newdata, type = "response")
    },
    varimp = function(object, ...) {
      pchisq(coef(object)^2 / diag(vcov(object)), 1)
    }
  )
}

## Add to MachineShop for full compatibility
environment(LogisticModel) <- asNamespace("MachineShop")

}
\seealso{
\code{\link{modelinfo}}
}
