% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelList.R
\name{ModelList}
\alias{ModelList}
\alias{MLModelList}
\title{Model List}
\usage{
ModelList(...)
}
\arguments{
\item{...}{\link[=models]{model} functions, function names, calls, or vectors
of these.}
}
\value{
\code{MLModelList} class object that inherits from \code{list}.
}
\description{
Create a list of models.
}
\examples{
ModelList(GBMModel, GLMNetModel, RandomForestModel)

}
\seealso{
\code{\link{tune}}, \code{\link{SelectedModel}}
}
