% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ML_NaiveBayesModel.R
\name{NaiveBayesModel}
\alias{NaiveBayesModel}
\title{Naive Bayes Classifier Model}
\usage{
NaiveBayesModel(laplace = 0)
}
\arguments{
\item{laplace}{positive numeric controlling Laplace smoothing.}
}
\value{
\code{MLModel} class object.
}
\description{
Computes the conditional a-posterior probabilities of a categorical class
variable given independent predictor variables using Bayes rule.
}
\details{
\describe{
  \item{Response types:}{\code{factor}}
}

Further model details can be found in the source link below.
}
\examples{
\donttest{
## Requires prior installation of suggested package e1071 to run

fit(Species ~ ., data = iris, model = NaiveBayesModel)
}

}
\seealso{
\code{\link[e1071]{naiveBayes}}, \code{\link{fit}},
\code{\link{resample}}
}
