% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.VarImp}
\alias{plot.VarImp}
\title{Variable Importance Plot}
\usage{
\method{plot}{VarImp}(x, n = NULL, ...)
}
\arguments{
\item{x}{variable importance object as returned by \code{\link{varimp}}.}

\item{n}{number of most important variables to include in the plot
(default: all).}

\item{...}{arguments to be passed to other methods.}
}
\description{
Plot measures of the relative importance of predictors in a model.
}
\seealso{
\code{\link{varimp}}
}
