% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ML_POLRModel.R
\name{POLRModel}
\alias{POLRModel}
\title{Ordered Logistic or Probit Regression Model}
\usage{
POLRModel(method = c("logistic", "probit", "loglog", "cloglog",
  "cauchit"))
}
\arguments{
\item{method}{logistic or probit or (complementary) log-log or cauchit
(corresponding to a Cauchy latent variable).}
}
\value{
\code{MLModel} class object.
}
\description{
Fit a logistic or probit regression model to an ordered factor response.
}
\details{
\describe{
\item{Response Types:}{\code{ordered}}
}

Further model details can be found in the source link below.
}
\examples{
df <- within(ICHomes,
             sale_amount <- cut(sale_amount,
             breaks = c(0, 100, 200, 300, 1000) * 1000,
             ordered = TRUE))
fit(sale_amount ~ ., data = df, model = POLRModel())

}
\seealso{
\code{\link[MASS]{polr}}, \code{\link{fit}}, \code{\link{resample}},
\code{\link{tune}}
}
