% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLRandomForestModel.R
\name{RandomForestModel}
\alias{RandomForestModel}
\title{Random Forest Model}
\usage{
RandomForestModel(ntree = NULL, mtry = NULL, replace = NULL,
  nodesize = NULL, maxnodes = NULL)
}
\arguments{
\item{ntree}{number of trees to grow.}

\item{mtry}{number of variables randomly sampled as candidates at each split.}

\item{replace}{should sampling of cases be done with or without replacement?}

\item{nodesize}{minimum size of terminal nodes.}

\item{maxnodes}{maximum number of terminal nodes trees in the forest can
have.}
}
\value{
MLModel class object.
}
\description{
Implementation of Breiman's random forest algorithm (based on Breiman and
Cutler's original Fortran code) for classification and regression.
}
\details{
\describe{
\item{Response Types:}{\code{factor}, \code{numeric}}
}

Default values for the \code{NULL} arguments and further model details can be
found in the source link below.
}
\seealso{
\code{\link[randomForest]{randomForest}}, \code{\link{fit}},
\code{\link{resample}}, \code{\link{tune}}
}
