% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{Resamples-class}
\alias{Resamples-class}
\alias{Resamples}
\title{Resamples Class Contructor}
\arguments{
\item{...}{named or unnamed resample output from one or more models.}
}
\value{
Resamples class object.
}
\description{
Create an object of resampled performance metrics from one or more models.
}
\seealso{
\code{\link{resample}}, \code{\link{plot}}, \code{\link{summary}}
}
