% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{Resamples}
\alias{Resamples}
\title{Resamples Class Contructor}
\usage{
Resamples(..., method = NULL)
}
\arguments{
\item{...}{named or unnamed resample output from one or more models.}

\item{method}{character string indicating the type of resampling method.
This need only be specified if the supplied output is not a Resamples object.}
}
\value{
Resamples class object.
}
\description{
Create an object of resampled performance metrics from one or more models.
}
\seealso{
\code{\link{resample}}, \code{\link{plot}}, \code{\link{summary}}
}
