% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfe.R
\name{rfe}
\alias{rfe}
\alias{rfe.formula}
\alias{rfe.matrix}
\alias{rfe.ModelFrame}
\alias{rfe.recipe}
\alias{rfe.MLModel}
\alias{rfe.MLModelFunction}
\title{Recursive Feature Elimination}
\usage{
rfe(x, ...)

\method{rfe}{formula}(
  x,
  data,
  model,
  control = MachineShop::settings("control"),
  props = 4,
  sizes = NULL,
  recompute = FALSE,
  optimize = c("global", "local"),
  samples = c(rfe = 1, varimp = 1),
  metrics = NULL,
  stat = "base::mean",
  ...
)

\method{rfe}{matrix}(
  x,
  y,
  model,
  control = MachineShop::settings("control"),
  props = 4,
  sizes = NULL,
  recompute = FALSE,
  optimize = c("global", "local"),
  samples = c(rfe = 1, varimp = 1),
  metrics = NULL,
  stat = "base::mean",
  ...
)

\method{rfe}{ModelFrame}(
  x,
  model,
  control = MachineShop::settings("control"),
  props = 4,
  sizes = NULL,
  recompute = FALSE,
  optimize = c("global", "local"),
  samples = c(rfe = 1, varimp = 1),
  metrics = NULL,
  stat = "base::mean",
  ...
)

\method{rfe}{recipe}(
  x,
  model,
  control = MachineShop::settings("control"),
  props = 4,
  sizes = NULL,
  recompute = FALSE,
  optimize = c("global", "local"),
  samples = c(rfe = 1, varimp = 1),
  metrics = NULL,
  stat = "base::mean",
  ...
)

\method{rfe}{MLModel}(x, ...)

\method{rfe}{MLModelFunction}(x, ...)
}
\arguments{
\item{x}{\link[=inputs]{input} specifying a relationship between model
predictor and response variables.  Alternatively, a \link[=models]{model}
function or object may be given first followed by the input specification.}

\item{...}{arguments passed to other methods.}

\item{data}{\link[=data.frame]{data frame} containing observed predictors and
outcomes.}

\item{model}{\link[=models]{model} function, function name, or object;
ignored and can be omitted when fitting
\link[=ModeledInput]{modeled inputs}.}

\item{control}{\link[=controls]{control} function, function name, or object
defining the resampling method to be employed.}

\item{props}{numeric vector of the proportions of most important predictor
variables to retain in fitted models or an integer number of equal spaced
proportions to generate automatically; ignored if \code{sizes} are given.}

\item{sizes}{integer vector of the set sizes of most important predictor
variables to retain.}

\item{recompute}{logical indicating whether to recompute variable importance
after eliminating each set of variables.}

\item{optimize}{character string specifying a search through all \code{props}
to identify the globally optimal model (\code{"global"}) or a search that
stops after identifying the first locally optimal model (\code{"local"}).}

\item{samples}{numeric vector or list giving the number of permutation
samples for each of the \code{rfe} and \code{\link{varimp}} algorithms.
One or both of the values may be specified as named arguments or in the
order in which their defaults appear.  Larger numbers of samples decrease
variability in estimated model performances and variable importances at the
expense of increased computation time.  Samples are more expensive
computationally for \code{rfe} than for \code{varimp}.}

\item{metrics}{\link[=metrics]{metric} function, function name, or vector of
these with which to calculate performance.  If not specified, default
metrics defined in the \link{performance} functions are used.}

\item{stat}{function or character string naming a function to compute a
summary statistic on resampled metric values and permuted samples.}

\item{y}{response variable.}
}
\value{
A data frame with columns for the numbers of predictor variables
retained (size), their names (terms), logical indicators to identify the
optimal model (optimal), and associated predictive performances
(performance).
}
\description{
A wrapper method of backward feature selection in which a given model is fit
to nested subsets of most important predictor variables in order to select
the subset whose resampled predictive performance is optimal.
}
\examples{
\donttest{
## Requires prior installation of suggested package gbm to run

rfe(sale_amount ~ ., data = ICHomes, model = GBMModel)
}

}
\seealso{
\code{\link{varimp}}
}
