% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary}
\alias{summary}
\alias{summary.Performance}
\alias{summary.Resamples}
\alias{summary.MLModelTune}
\alias{summary.Confusion}
\alias{summary.ConfusionMatrix}
\alias{summary.Curves}
\title{Model Performance Summary}
\usage{
\method{summary}{Performance}(object,
  stats = MachineShop::settings("stats.Resamples"), na.rm = TRUE, ...)

\method{summary}{Resamples}(object,
  stats = MachineShop::settings("stats.Resamples"), na.rm = TRUE, ...)

\method{summary}{MLModelTune}(object,
  stats = MachineShop::settings("stats.Resamples"), na.rm = TRUE, ...)

\method{summary}{Confusion}(object, ...)

\method{summary}{ConfusionMatrix}(object, ...)

\method{summary}{Curves}(object,
  stat = MachineShop::settings("stat.Curves"), ...)
}
\arguments{
\item{object}{\link{confusion}, \link{lift}, model \link{tune},
\link{performance}, \link[=performance_curve]{performance curve}, or
\link{resample} result.}

\item{stats}{function, function name, or vector of these with which to
compute summary statistics.}

\item{na.rm}{logical indicating whether to exclude missing values.}

\item{...}{arguments passed to other methods.}

\item{stat}{function or character string naming a function to compute a
summary statistic at each cutoff value of resampled metrics in \code{Curves},
or \code{NULL} for resample-specific metrics.}
}
\value{
array with summmary statistics in the second dimension, metrics in
the first for single models, and models and metrics in the first and third,
respectively, for multiple models.
}
\description{
Summary statistics for resampled model performance metrics.
}
\examples{
## Factor response example

fo <- Species ~ .
control <- CVControl()

gbmres1 <- resample(fo, iris, GBMModel(n.trees = 25), control)
gbmres2 <- resample(fo, iris, GBMModel(n.trees = 50), control)
gbmres3 <- resample(fo, iris, GBMModel(n.trees = 100), control)
summary(gbmres3)

res <- Resamples(GBM1 = gbmres1, GBM2 = gbmres2, GBM3 = gbmres3)
summary(res)

}
