% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary}
\alias{summary}
\alias{summary.Resamples}
\title{Model Performance Summary}
\usage{
\method{summary}{Resamples}(object, stats = c(Mean = mean, Median =
  median, SD = sd, Min = min, Max = max), na.rm = TRUE, ...)
}
\arguments{
\item{object}{object of resampled metrics as returned by
\code{\link{resample}}.}

\item{stats}{list of named functions to include in the calculation of summary
statistics.  Supplied functions should contain a \code{na.rm} argument in
their definitions.}

\item{na.rm}{logical indicating whether to exclude missing values.}

\item{...}{arguments to be passed to other methods.}
}
\description{
Summary statistics for resampled model performance metrics.
}
\seealso{
\code{\link{resample}}
}
