% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{SurvMatrix}
\alias{SurvMatrix}
\alias{SurvEvents}
\alias{SurvProbs}
\title{SurvMatrix Class Constructor}
\usage{
SurvEvents(object = numeric(), times = NULL)

SurvProbs(object = numeric(), times = NULL)
}
\arguments{
\item{object}{matrix, or object that can be converted to one, of predicted
survival events or probabilities with columns and rows representing
prediction times and cases, respectively.}

\item{times}{numeric vector of the survival prediction times.}
}
\value{
Object that is of the same class as the constructor name and inherits
from \code{SurvMatrix}.  Examples of these objects are the predicted survival
events and probabilities returned by the \code{predict} function.
}
\description{
Create an object of predicted survival events or probabilites for use with
metrics provided by the \pkg{MachineShop} package.
}
\seealso{
\code{\link{metrics}}, \code{\link{predict}}
}
