% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLModel.R
\name{unMLModelFit}
\alias{unMLModelFit}
\title{Revert an MLModelFit Object}
\usage{
unMLModelFit(object)
}
\arguments{
\item{object}{model \link{fit} result.}
}
\value{
The supplied object with its \code{MLModelFit} classes and fields
removed.
}
\description{
Function to revert an \code{MLModelFit} object to its original class.
}
