% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varimp.R
\name{varimp}
\alias{varimp}
\title{Variable Importance}
\usage{
varimp(object, scale = TRUE, ...)
}
\arguments{
\item{object}{model \link{fit} result.}

\item{scale}{logical indicating whether importance measures should be scaled
to range from 0 to 100.}

\item{...}{arguments passed to model-specific variable importance functions.}
}
\value{
\code{VarImp} class object.
}
\description{
Calculate measures of the relative importance of predictors in a model.
}
\examples{
\donttest{
## Requires prior installation of suggested package gbm to run

## Survival response example
library(survival)

gbm_fit <- fit(Surv(time, status) ~ ., data = veteran, model = GBMModel)
(vi <- varimp(gbm_fit))
plot(vi)
}

}
\seealso{
\code{\link{plot}}
}
