% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLModel.R
\name{as.MLModel}
\alias{as.MLModel}
\title{Coerce to an MLModel}
\usage{
as.MLModel(x, ...)
}
\arguments{
\item{x}{model \link{fit} result.}

\item{...}{arguments passed to other methods.}
}
\value{
\code{MLModel} class object.
}
\description{
Function to coerce an \code{MLModelFit} object to an \code{MLModel}.
}
