% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coerce.R
\name{as.MLModel}
\alias{as.MLModel}
\alias{as.MLModel.MLModelFit}
\alias{as.MLModel.ModelSpecification}
\alias{as.MLModel.model_spec}
\title{Coerce to an MLModel}
\usage{
as.MLModel(x, ...)

\method{as.MLModel}{MLModelFit}(x, ...)

\method{as.MLModel}{ModelSpecification}(x, ...)

\method{as.MLModel}{model_spec}(x, ...)
}
\arguments{
\item{x}{model \link{fit} result, \pkg{MachineShop}
\link[=ModelSpecification]{model specification}, or
\pkg{parsnip} \link[parsnip:model_spec]{model specification}.}

\item{...}{arguments passed to other methods.}
}
\value{
\code{MLModel} class object.
}
\description{
Function to coerce an object to \code{MLModel}.
}
\seealso{
\code{\link{ParsnipModel}}
}
