% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c.R
\name{c}
\alias{c}
\alias{combine}
\alias{c.Calibration}
\alias{c.ConfusionList}
\alias{c.ConfusionMatrix}
\alias{c.Curves}
\alias{c.Lift}
\alias{c.ListOf}
\alias{c.Resamples}
\title{Combine MachineShop Objects}
\usage{
\method{c}{Calibration}(...)

\method{c}{ConfusionList}(...)

\method{c}{ConfusionMatrix}(...)

\method{c}{Curves}(...)

\method{c}{Lift}(...)

\method{c}{ListOf}(...)

\method{c}{Resamples}(...)
}
\arguments{
\item{...}{named or unnamed \link{calibration}, \link{confusion},
\link[=curves]{performance curve}, \link{lift}, or \link{resample} results.
Curves must have been generated with the same performance \link{metrics} and
resamples with the same resampling \link[=controls]{control}.}
}
\value{
Object of the same class as the arguments.
}
\description{
Combine one or more \pkg{MachineShop} objects of the same class.
}
