% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/calibration.R
\name{calibration}
\alias{calibration}
\alias{Calibration}
\title{Model Calibration}
\usage{
Calibration(...)

calibration(x, y = NULL, breaks = 10, dist = NULL, ...)
}
\arguments{
\item{...}{named or unnamed \code{calibration} output to combine together
with the \code{Calibration} constructor.}

\item{x}{observed responses or \code{Resamples} object of observed and
predicted responses.}

\item{y}{predicted responses.}

\item{breaks}{value defining the response variable bins within which to
calculate observed mean values.  May be specified as a number of bins, a
vector of breakpoints, or \code{NULL} to fit smooth curves with splines for
survival responses and loess for others.}

\item{dist}{character string specifying a distribution with which to estimate
observed survival means.  Possible values are \code{"empirical"} (default)
for the Kaplan-Meier estimator, \code{"exponential"}, \code{"extreme"},
\code{"gaussian"}, \code{"loggaussian"}, \code{"logistic"},
\code{"loglogistic"}, \code{"lognormal"}, \code{"rayleigh"}, \code{"t"}, or
\code{"weibull"}.}
}
\value{
\code{Calibration} class object that inherits from \code{data.frame}.
}
\description{
Calculate calibration estimates from observed and predicted responses.
}
\examples{
library(survival)
library(MASS)

res <- resample(Surv(time, status != 2) ~ sex + age + year + thickness + ulcer,
                data = Melanoma, model = GBMModel,
                control = CVControl(times = 365 * c(2, 5, 10)))
cal <- calibration(res)
plot(cal)

}
\seealso{
\code{\link{response}}, \code{\link{predict}},
\code{\link{resample}}, \code{\link{plot}}
}
