% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confusion.R
\name{confusion}
\alias{confusion}
\title{Confusion Matrix}
\usage{
confusion(x, ...)
}
\arguments{
\item{x}{\code{Resamples} object.}

\item{...}{arguments passed to other methods.}
}
\value{
\code{ConfusionResamples} class object that inherits from
\code{list}.
}
\description{
Calculate confusion matrices of cross-classified predicted and observed
responses.
}
\details{
Responses are summed directly over predicted class probabilities for the
cross-classification and will thus appear as decimal numbers that can be
interpreted as expected counts.
}
\examples{
res <- resample(Species ~ ., data = iris, model = GBMModel)
confusion(res)

}
\seealso{
\code{\link{resample}}
}
