% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditions.R
\name{deprecated}
\alias{deprecated}
\alias{Grid}
\alias{ModeledInput}
\title{Deprecated Functions}
\usage{
Grid(...)

ModeledInput(...)
}
\arguments{
\item{...}{arguments passed to non-deprecated equivalent.}
}
\description{
Functions that have been deprecated and will be removed in a future version
of the package.
}
\details{
Use \code{\link[=TuningGrid]{TuningGrid()}} instead of \code{Grid()}.

Use \code{\link[=ModelSpecification]{ModelSpecification()}} instead of
\code{ModeledInput()}.
}
