% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{deprecated}
\alias{deprecated}
\alias{SelectedFormula}
\alias{SelectedMatrix}
\alias{SelectedModelFrame}
\alias{SelectedRecipe}
\alias{tune}
\alias{TunedRecipe}
\title{Deprecated Functions}
\usage{
SelectedFormula(...)

SelectedMatrix(...)

SelectedModelFrame(...)

SelectedRecipe(...)

tune(...)

TunedRecipe(...)
}
\arguments{
\item{...}{arguments passed to non-deprecated equivalent.}
}
\description{
Functions that have been deprecated and will be removed in the future.
}
\details{
Use \code{\link[=SelectedInput]{SelectedInput()}} instead of
\code{SelectedFormula()}.

Use \code{\link[=SelectedInput]{SelectedInput()}} instead of
\code{SelectedMatrix()}.

Use \code{\link[=SelectedInput]{SelectedInput()}} instead of
\code{SelectedModelFrame()}.

Use \code{\link[=SelectedInput]{SelectedInput()}} instead of
\code{SelectedRecipe()}.

Use \code{\link[=TunedInput]{TunedInput()}} instead of \code{TunedRecipe()}.
}
