% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{tune}
\alias{tune}
\title{Deprecated Functions}
\usage{
tune(...)
}
\arguments{
\item{...}{arguments passed to non-deprecated equivalent.}
}
\description{
Functions that have been deprecated and will be removed in the future.
}
