% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response.R
\name{DiscreteVector}
\alias{DiscreteVector}
\alias{BinomialVector}
\alias{NegBinomialVector}
\alias{PoissonVector}
\title{DiscreteVector Class Constructors}
\usage{
BinomialVector(object, size)

DiscreteVector(object, min = -Inf, max = Inf)

NegBinomialVector(object)

PoissonVector(object)
}
\arguments{
\item{object}{numeric vector.}

\item{size}{number of binomial trials.}

\item{min, max}{minimum and maximum bounds for discrete numbers.}
}
\value{
\code{DiscreteVector} object class that inherits from \code{integer}
or object that is of the same class as the constructor name and inherits
from \code{DiscreteVector}.
}
\description{
Create an integer vector of binomial counts, discrete numbers, negative
binomial counts, or Poisson counts.
}
\examples{
BinomialVector(rbinom(25, 10, 0.5), size = 10)
PoissonVector(rpois(25, 10))

}
