% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract}
\alias{extract}
\alias{[,DiscreteVector,ANY,missing,missing-method}
\alias{[,RecipeGrid,ANY,ANY,ANY-method}
\alias{[,Resamples,ANY,ANY,ANY-method}
\alias{[,Resamples,ANY,missing,ANY-method}
\alias{[,Resamples,missing,missing,ANY-method}
\alias{[,SurvMatrix,ANY,ANY,ANY-method}
\alias{+,SurvMatrix,SurvMatrix-method}
\title{Extract Elements of an Object}
\usage{
\S4method{[}{DiscreteVector,ANY,missing,missing}(x, i)

\S4method{[}{RecipeGrid,ANY,ANY,ANY}(x, i, j, ..., drop = FALSE)

\S4method{[}{Resamples,ANY,ANY,ANY}(x, i, j, ..., drop = FALSE)

\S4method{[}{Resamples,ANY,missing,ANY}(x, i, j, ..., drop = FALSE)

\S4method{[}{Resamples,missing,missing,ANY}(x, i, j, ..., drop = FALSE)

\S4method{[}{SurvMatrix,ANY,ANY,ANY}(x, i, j, ..., drop = FALSE)

\S4method{+}{SurvMatrix,SurvMatrix}(e1, e2)
}
\arguments{
\item{x}{object from which to extract elements.}

\item{i, j, ...}{indices specifying elements to extract.}

\item{drop}{logical indicating that the result be returned as an object
coerced to the lowest dimension possible if \code{TRUE} or
with the original dimensions and class otherwise.}

\item{e1, e2}{objects.}
}
\description{
Operators acting on data structures to extract elements.
}
