% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\docType{methods}
\name{fit}
\alias{fit}
\alias{fit,MLModel,data.frame-method}
\alias{fit,MLModel,formula-method}
\alias{fit,MLModel,recipe-method}
\title{Model Fitting}
\usage{
fit(object, x, ...)

\S4method{fit}{MLModel,data.frame}(object, x, ...)

\S4method{fit}{MLModel,formula}(object, x, data, ...)

\S4method{fit}{MLModel,recipe}(object, x, ...)
}
\arguments{
\item{object}{prediction model object.}

\item{x}{defined relationship between model predictors and an outcome.  May
be a model.frame (data.frame) containing a formula, data, and optionally case
weights; a formula; or a recipe.}

\item{...}{further arguments passed to other methods.}

\item{data}{data frame containing observed predictors and outcomes.}
}
\description{
Fit a model to estimate its parameters from a data set.
}
\seealso{
\code{\link[stats]{model.frame}}, \code{\link[recipes]{recipe}},
\code{\link{predict}}, \code{\link{varimp}}
}
