% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLGLMNetModel.R
\name{GLMNetModel}
\alias{GLMNetModel}
\title{GLM Lasso or Elasticnet Model}
\usage{
GLMNetModel(family = NULL, alpha = NULL, lambda = NULL,
  standardize = NULL, thresh = NULL, maxit = NULL,
  type.gaussian = NULL, type.logistic = NULL,
  type.multinomial = NULL)
}
\arguments{
\item{family}{response type.}

\item{alpha}{elasticnet mixing parameter.}

\item{lambda}{regularization parameter.}

\item{standardize}{logical flag for predictor variable standardization, prior
to model fitting.}

\item{thresh}{convergence threshold for coordinate descent.}

\item{maxit}{maximum number of passes over the data for all lambda values.}

\item{type.gaussian}{algorithm type for guassian models.}

\item{type.logistic}{algorithm type for logistic models.}

\item{type.multinomial}{algorithm type for multinomial models.}
}
\description{
Fit a generalized linear model via penalized maximum likelihood.
}
\details{
\describe{
\item{Response Types:}{\code{factor}, \code{numeric}, \code{Surv}}
}

Default values for the \code{NULL} arguments and further model
details can be found in the source link below.
}
\seealso{
\code{\link[glmnet]{glmnet}}
}
