% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{Grid}
\alias{Grid}
\title{Tuning Grid Control}
\usage{
Grid(length = 3, random = FALSE)
}
\arguments{
\item{length}{number of values to be generated for each model parameter in
the tuning grid.}

\item{random}{number of unique grid points to sample at random, \code{Inf}
for all random points, or \code{FALSE} for all fixed points.}
}
\value{
\code{Grid} class object.
}
\description{
Defines the control parameters for a tuning grid.
}
\seealso{
\code{\link{tune}}
}
