% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lift.R
\name{lift}
\alias{lift}
\title{Model Lift}
\usage{
lift(x, y = NULL, ...)
}
\arguments{
\item{x}{observed responses or \code{Resamples} object of observed and
predicted responses.}

\item{y}{predicted responses.}

\item{...}{arguments passed to other methods.}
}
\value{
\code{Lift} class object that inherits from \code{data.frame}.
}
\description{
Calculate lift estimates from observed and predicted responses.
}
\examples{
library(MASS)

res <- resample(type ~ ., data = Pima.tr, model = GBMModel)
(lf <- lift(res))
plot(lf)

}
\seealso{
\code{\link{response}}, \code{\link{predict}},
\code{\link{resample}}, \code{\link{plot}}
}
