% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lift.R
\name{lift}
\alias{lift}
\title{Model Lift}
\usage{
lift(x, ...)
}
\arguments{
\item{x}{\code{Resamples} object.}

\item{...}{arguments passed to other methods.}
}
\value{
\code{LiftResamples} class object that inherits from
\code{data.frame}.
}
\description{
Calculate lift estimates from observed and resampled response variable
values.
}
\examples{
library(MASS)

res <- resample(type ~ ., data = Pima.tr, model = GBMModel)
(lf <- lift(res))
plot(lf)

}
\seealso{
\code{\link{resample}}, \code{\link{plot}}
}
