% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ML_LMModel.R
\name{LMModel}
\alias{LMModel}
\title{Linear Models}
\usage{
LMModel()
}
\value{
\code{MLModel} class object.
}
\description{
Fits linear models.
}
\details{
\describe{
  \item{Response Types:}{\code{factor}, \code{matrix}, \code{numeric}}
}

Further model details can be found in the source link below.
}
\examples{
fit(sale_amount ~ ., data = ICHomes, model = LMModel)

}
\seealso{
\code{\link[stats]{lm}}, \code{\link{fit}}, \code{\link{resample}}
}
