% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metricinfo.R
\name{metricinfo}
\alias{metricinfo}
\title{Display Performance Metric Information}
\usage{
metricinfo(...)
}
\arguments{
\item{...}{one or more metric function names, observed response, observed and
predicted responses, or a \code{Resamples} object.  If none are specified,
information is returned on all available metrics by default.}
}
\value{
List of named metrics available for the supplied arguments and
containing a descriptive \code{"label"}, the functions' \code{"arguments"},
and supported response variable \code{"types"}.
}
\description{
Display information about metrics provided by the \pkg{MachineShop} package.
}
\examples{
## All metrics
metricinfo()

## Metrics by observed and predicted response types
names(metricinfo(factor(0)))
names(metricinfo(factor(0), factor(0)))
names(metricinfo(factor(0), matrix(0)))
names(metricinfo(factor(0), numeric(0)))

}
\seealso{
\code{\link{metrics}}, \code{\link{Resamples}}
}
