% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics-factor.R, R/metrics-numeric.R
\name{metrics}
\alias{metrics}
\alias{accuracy}
\alias{auc}
\alias{brier}
\alias{cindex}
\alias{cross_entropy}
\alias{f_score}
\alias{fnr}
\alias{fpr}
\alias{kappa2}
\alias{npv}
\alias{ppv}
\alias{pr_auc}
\alias{precision}
\alias{recall}
\alias{roc_auc}
\alias{roc_index}
\alias{rpp}
\alias{sensitivity}
\alias{specificity}
\alias{tnr}
\alias{tpr}
\alias{weighted_kappa2}
\alias{gini}
\alias{mae}
\alias{mse}
\alias{msle}
\alias{r2}
\alias{rmse}
\alias{rmsle}
\title{Performance Metrics}
\usage{
accuracy(observed, predicted = NULL, cutoff = 0.5, times = numeric(),
  ...)

auc(observed, predicted = NULL, metrics = c(MachineShop::tpr,
  MachineShop::fpr), times = numeric(), stat = base::mean, ...)

brier(observed, predicted = NULL, times = numeric(), ...)

cindex(observed, predicted = NULL, ...)

cross_entropy(observed, predicted = NULL, ...)

f_score(observed, predicted = NULL, cutoff = 0.5, times = numeric(),
  beta = 1, ...)

fnr(observed, predicted = NULL, cutoff = 0.5, times = numeric(), ...)

fpr(observed, predicted = NULL, cutoff = 0.5, times = numeric(), ...)

kappa2(observed, predicted = NULL, cutoff = 0.5, times = numeric(),
  ...)

npv(observed, predicted = NULL, cutoff = 0.5, times = numeric(), ...)

ppv(observed, predicted = NULL, cutoff = 0.5, times = numeric(), ...)

pr_auc(observed, predicted = NULL, times = numeric(), ...)

precision(observed, predicted = NULL, cutoff = 0.5,
  times = numeric(), ...)

recall(observed, predicted = NULL, cutoff = 0.5, times = numeric(),
  ...)

roc_auc(observed, predicted = NULL, times = numeric(), ...)

roc_index(observed, predicted = NULL, cutoff = 0.5,
  times = numeric(), f = function(sens, spec) sens + spec, ...)

rpp(observed, predicted = NULL, cutoff = 0.5, times = numeric(), ...)

sensitivity(observed, predicted = NULL, cutoff = 0.5,
  times = numeric(), ...)

specificity(observed, predicted = NULL, cutoff = 0.5,
  times = numeric(), ...)

tnr(observed, predicted = NULL, cutoff = 0.5, times = numeric(), ...)

tpr(observed, predicted = NULL, cutoff = 0.5, times = numeric(), ...)

weighted_kappa2(observed, predicted = NULL, power = 1, ...)

gini(observed, predicted = NULL, ...)

mae(observed, predicted = NULL, ...)

mse(observed, predicted = NULL, ...)

msle(observed, predicted = NULL, ...)

r2(observed, predicted = NULL, ...)

rmse(observed, predicted = NULL, ...)

rmsle(observed, predicted = NULL, ...)
}
\arguments{
\item{observed}{observed responses, \code{\link{Curves}} object, or
\code{\link[=confusion]{ConfusionMatrix}} of observed and predicted
responses.}

\item{predicted}{predicted responses.}

\item{cutoff}{threshold above which binary factor probabilities are
classified as events and below which survival probabilities are classified.}

\item{times}{numeric vector of follow-up times at which survival
probabilities were predicted.}

\item{...}{arguments passed to or from other methods.}

\item{metrics}{list of two performance metrics for the calculation [default:
ROC metrics].}

\item{stat}{function to compute a summary statistic at each cutoff value of
resampled metrics in \code{Curves}, or \code{NULL} for resample-specific
metrics.}

\item{beta}{relative importance of recall to precision in the calculation of
\code{f_score} [default: F1 score].}

\item{f}{function to calculate a desired sensitivity-specificity tradeoff.}

\item{power}{power to which positional distances of off-diagonals from the
main diagonal in confusion matrices are raised to calculate
\code{weighted_kappa2}.}
}
\description{
Compute measures of agreement between observed and predicted responses.
}
\seealso{
\code{\link{metricinfo}}, \code{\link{confusion}},
\code{\link{performance}}, \code{\link{performance_curve}}
}
