% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{accuracy}
\alias{accuracy}
\alias{metrics}
\alias{brier}
\alias{cindex}
\alias{cross_entropy}
\alias{kappa}
\alias{mae}
\alias{mse}
\alias{pr_auc}
\alias{r2}
\alias{rmse}
\alias{roc_auc}
\alias{roc_index}
\alias{sensitivity}
\alias{specificity}
\alias{weighted_kappa}
\title{Performance Metrics}
\usage{
accuracy(observed, predicted, cutoff = 0.5, ...)

brier(observed, predicted, times = numeric(), ...)

cindex(observed, predicted, ...)

cross_entropy(observed, predicted, ...)

kappa(observed, predicted, cutoff = 0.5, ...)

mae(observed, predicted, ...)

mse(observed, predicted, ...)

pr_auc(observed, predicted, ...)

r2(observed, predicted, ...)

rmse(observed, predicted, ...)

roc_auc(observed, predicted, times = numeric(), ...)

roc_index(observed, predicted, cutoff = 0.5,
  cutoff_index = function(sens, spec) sens + spec, ...)

sensitivity(observed, predicted, cutoff = 0.5, ...)

specificity(observed, predicted, cutoff = 0.5, ...)

weighted_kappa(observed, predicted, ...)
}
\arguments{
\item{observed}{observed responses.}

\item{predicted}{predicted responses.}

\item{cutoff}{threshold above which probabilities are classified as success
for binary responses.}

\item{...}{arguments passed to or from other methods.}

\item{times}{numeric vector of follow-up times at which survival events
were predicted.}

\item{cutoff_index}{function to calculate a desired sensitivity-specificity
tradeoff.}
}
\description{
Compute measures of agreement between observed and predicted responses.
}
