% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{MLControl-class}
\alias{MLControl-class}
\alias{initialize,MLControl-method}
\alias{BootControl}
\alias{CVControl}
\alias{OOBControl}
\title{Resampling Classes and Methods}
\usage{
\S4method{initialize}{MLControl}(.Object, summary = modelmetrics,
  cutoff = 0.5, cutoff.index = function(sens, spec) sens + spec,
  survtimes = numeric(), ...)

BootControl(number = 25, ...)

CVControl(folds = 10, repeats = 1, ...)

OOBControl(number = 25, ...)
}
\arguments{
\item{.Object}{class object being initialized.}

\item{summary}{function to compute model performance metrics.}

\item{cutoff}{threshold above which probabilities are classified as success
for factor outcomes and which expected values are rounded for integer
outcomes.}

\item{cutoff.index}{function to calculate a desired sensitivity-specificity
tradeoff.}

\item{survtimes}{numeric vector of follow-up times at which to predict
survival events.}

\item{...}{arguments to be passed to or from other methods.}

\item{number}{number of bootstrap resamples.}

\item{folds}{number of cross-validation folds (K).}

\item{repeats}{number of cross-validation replicates.}
}
\value{
MLControl class object.
}
\description{
Resampling Classes and Methods

The base MLControl constructor initializes a set of parameters that are common
to all resampling methods.

\code{BootControl} constructs an MLControl object for simple bootstrap
resampling in which models are estimated with bootstrap resampled data sets
and evaluated on the full data set.

\code{CVControl} constructs an MLControl object for repeated K-fold
cross-validation.  In this procedure, the data set is repeatedly partitioned
into K-folds.  Evaluation is performed on each of the K folds with models
estimated on the remaining folds.

\code{OOBControl} constructs an MLControl object for out-of-bag bootstrap
resampling in which models are estimated on bootstrap resampled data sets and
evaluated on unsampled cases.
}
\section{Slots}{

\describe{
\item{\code{summary}}{function to compute model performance metrics.}

\item{\code{cutoff}}{threshold above which probabilities are classified as success
for factor outcomes and which expected values are rounded for integer
outcomes.}

\item{\code{cutoff.index}}{function to calculate a desired sensitivity-specificity
tradeoff.}

\item{\code{survtimes}}{numeric vector of follow-up times at which to predict
survival events.}
}}

\seealso{
\code{\link{resample}}, \code{\link{modelmetrics}}
}
