% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{MLModel}
\alias{MLModel}
\title{MLModel Class Constructor}
\usage{
MLModel(name = "MLModel", packages = character(0),
  types = character(0), params = list(), nvars = function(data) NULL,
  fit = function(formula, data, weights, ...) stop("no fit function"),
  predict = function(object, newdata, times, ...)
  stop("no predict function"), varimp = function(object, ...) NULL)
}
\arguments{
\item{name}{character string name for the instantiated MLModel object.}

\item{packages}{character vector of packages required by the object.}

\item{types}{character vector of response variable types on which the model
can be fit.}

\item{params}{list of user-specified model parameters.}

\item{nvars}{function to return the number of predictor variables for a
given model frame.}

\item{fit}{model fitting function.}

\item{predict}{model prediction function.}

\item{varimp}{variable importance function.}
}
\description{
MLModel Class Constructor
}
