% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{MLModel}
\alias{MLModel}
\title{MLModel Class Constructor}
\usage{
MLModel(name = "MLModel", packages = character(0),
  types = character(0), params = list(), fit = function(...)
  stop("no fit function"), predict = function(...)
  stop("no predict function"), response = function(...)
  stop("no response function"), varimp = function(...)
  stop("no varimp function"))
}
\arguments{
\item{name}{character string name for the instantiated MLModel object.}

\item{packages}{character vector of packages required by the object.}

\item{types}{character vector of response variable types on which the model
can be fit.}

\item{params}{list of user-specified model parameters.}

\item{fit}{model fitting function.}

\item{predict}{model prediction function.}

\item{response}{function to extract the response variable from a model fit.}

\item{varimp}{variable importance function.}
}
\description{
MLModel Class Constructor
}
