% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{MLModelFit-class}
\alias{MLModelFit-class}
\title{Model Fit Class}
\description{
MLModelFit is the base class for model fit objects.  Direct calls to the
MLModelFit constructor are not necessary unless implementing custom models.
}
\section{Slots}{

\describe{
\item{\code{.packages}}{character vector of packages required by the object.}

\item{\code{.predict}}{model prediction function.}

\item{\code{.varimp}}{variable importance function.}

\item{\code{.y}}{observed response variable.}
}}

