% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelFrame.R
\name{ModelFrame}
\alias{ModelFrame}
\alias{ModelFrame.formula}
\alias{ModelFrame.matrix}
\title{ModelFrame Class}
\usage{
ModelFrame(x, ...)

\method{ModelFrame}{formula}(x, data, weights = NULL, strata = NULL,
  na.rm = TRUE, ...)

\method{ModelFrame}{matrix}(x, y, weights = NULL, strata = NULL,
  na.rm = TRUE, ...)
}
\arguments{
\item{x}{model \code{\link{formula}} or \code{matrix} of predictor variables.}

\item{...}{arguments passed to other methods.}

\item{data}{\code{data.frame} or an object that can be converted to one.}

\item{weights}{vector of case weights.}

\item{strata}{vector of stratification levels.}

\item{na.rm}{logical indicating whether to remove cases with \code{NA} values
for any of the model variables.}

\item{y}{response variable.}
}
\value{
\code{ModelFrame} class object that inherits from \code{data.frame}.
}
\description{
Class for storing a data frame, formula, and optionally weights for fitting
MLModels.
}
\examples{
mf <- ModelFrame(ncases / (ncases + ncontrols) ~ agegp + tobgp + alcgp,
                 data = esoph, weights = ncases + ncontrols)
gbmfit <- fit(mf, model = GBMModel)
varimp(gbmfit)

}
\seealso{
\code{\link{formula}}
}
