% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelFrame.R
\name{ModelFrame}
\alias{ModelFrame}
\alias{ModelFrame.formula}
\alias{ModelFrame.matrix}
\title{ModelFrame Class}
\usage{
ModelFrame(x, ...)

\method{ModelFrame}{formula}(x, data, strata = NULL, weights = NULL, na.rm = TRUE, ...)

\method{ModelFrame}{matrix}(
  x,
  y = NULL,
  offsets = NULL,
  strata = NULL,
  weights = NULL,
  na.rm = TRUE,
  ...
)
}
\arguments{
\item{x}{model \code{\link{formula}} or \code{\link{matrix}} of predictor
variables.  In the case of a formula, arguments \code{weights} and
\code{strata} are evaluated as expressions, whose objects are searched for
first in the accompanying \code{data} environment and, if not found there,
next in the calling environment.}

\item{...}{arguments passed to other methods.}

\item{data}{\link[=data.frame]{data frame} or an object that can be converted
to one.}

\item{strata}{vector of values to use in conducting stratified
\link{resample} estimation of model performance [default: none].}

\item{weights}{numeric vector of non-negative case weights for the \code{y}
response variable [default: equal weights].}

\item{na.rm}{logical indicating whether to remove cases with \code{NA} values
for any of the model variables.}

\item{y}{response variable.}

\item{offsets}{numeric vector, matrix, or data frame of values to be added
with a fixed coefficient of 1 to linear predictors in compatible regression
models.}
}
\value{
\code{ModelFrame} class object that inherits from \code{data.frame}.
}
\description{
Class for storing data, formulas, and other attributes for \pkg{MachineShop}
model fitting.
}
\examples{
\donttest{
## Requires prior installation of suggested package gbm to run

mf <- ModelFrame(ncases / (ncases + ncontrols) ~ agegp + tobgp + alcgp,
                 data = esoph, weights = ncases + ncontrols)
gbm_fit <- fit(mf, model = GBMModel)
varimp(gbm_fit)
}

}
\seealso{
\code{\link{fit}}, \code{\link{resample}}, \code{\link{response}},
\code{\link{SelectedInput}}
}
