% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelinfo.R
\name{modelinfo}
\alias{modelinfo}
\title{Display Model Information}
\usage{
modelinfo(...)
}
\arguments{
\item{...}{\code{MLModel} objects, constructor functions, constructor
function names, or supported responses for which to display information.  If
none are specified, information is returned on all available models by
default.}
}
\value{
List of named models containing a descriptive \code{"label"}, source
\code{"packages"}, supported response variable \code{"types"}, the
constructor \code{"arguments"}, whether there is a pre-defined \code{"grid"}
of tuning parameters, and whether a \code{"varimp"} function is implemented
for each.
}
\description{
Display information about models provided by the \pkg{MachineShop} package.
}
\examples{
## All models
modelinfo()

## Models by response types
names(modelinfo(factor(0)))
names(modelinfo(factor(0), numeric(0)))

}
\seealso{
\code{\link{fit}}, \code{\link{resample}}, \code{\link{tune}}
}
