% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelinfo.R
\name{modelinfo}
\alias{modelinfo}
\title{Display Model Information}
\usage{
modelinfo(...)
}
\arguments{
\item{...}{\code{MLModel} objects, constructor functions, character
strings naming constructor functions, or supported responses for which to
display information.  If none are specified, information is returned on all
available models by default.}
}
\value{
List of named models containing the \code{"label"}, required
\code{"packages"}, supported response variable \code{"types"}, the
constructor \code{"arguments"}, and whether a \code{"varimp"} function is
implemented for each.
}
\description{
Display information about models provided by the \pkg{MachineShop} package.
}
\examples{
## All models
modelinfo()

## Models by response types
modelinfo(factor(0))
modelinfo(factor(0), numeric(0))

}
