% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelmetrics.R
\name{modelmetrics}
\alias{modelmetrics}
\alias{modelmetrics.factor}
\alias{modelmetrics.matrix}
\alias{modelmetrics.numeric}
\alias{modelmetrics.Resamples}
\alias{modelmetrics.Surv}
\title{Model Performance Metrices}
\usage{
modelmetrics(x, ...)

\method{modelmetrics}{factor}(x, y, cutoff = 0.5,
  cutoff_index = function(sens, spec) sens + spec, ...)

\method{modelmetrics}{matrix}(x, y, ...)

\method{modelmetrics}{numeric}(x, y, ...)

\method{modelmetrics}{Resamples}(x, ...)

\method{modelmetrics}{Surv}(x, y, times = numeric(), ...)
}
\arguments{
\item{x}{observed responses or class of observed and predicted responses.}

\item{...}{arguments passed to or from other methods.}

\item{y}{predicted responses.}

\item{cutoff}{threshold above which probabilities are classified as success
for binary responses.}

\item{cutoff_index}{function to calculate a desired sensitivity-specificity
tradeoff.}

\item{times}{numeric vector of follow-up times at which survival events
were predicted.}
}
\value{
Numeric vector of model metrics.
}
\description{
Compute measures of model performance.
}
\examples{
## Survival response example
library(survival)
library(MASS)

fo <- Surv(time, status != 2) ~ sex + age + year + thickness + ulcer
gbmfit <- fit(fo, data = Melanoma, model = GBMModel)

obs <- response(fo, data = Melanoma)
pred <- predict(gbmfit, newdata = Melanoma, type = "prob")
modelmetrics(obs, pred)

}
\seealso{
\code{\link{predict}}, \code{\linkS4class{MLControl}}
}
