% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelmetrics.R
\name{modelmetrics}
\alias{modelmetrics}
\alias{modelmetrics.Resamples}
\alias{modelmetrics.factor}
\alias{modelmetrics.matrix}
\alias{modelmetrics.numeric}
\alias{modelmetrics.Surv}
\title{Model Performance Metrics}
\usage{
modelmetrics(x, ...)

\method{modelmetrics}{Resamples}(x, ..., na.rm = TRUE)

\method{modelmetrics}{factor}(x, y, metrics = c(Accuracy = accuracy,
  Kappa = kappa, WeightedKappa = weighted_kappa, Brier = brier,
  CrossEntropy = cross_entropy, ROCAUC = roc_auc, PRAUC = pr_auc,
  Sensitivity = sensitivity, Specificity = specificity, Index = roc_index),
  ...)

\method{modelmetrics}{matrix}(x, y, metrics = c(R2 = r2, RMSE = rmse, MAE
  = mae), ...)

\method{modelmetrics}{numeric}(x, y, metrics = c(R2 = r2, RMSE = rmse,
  MAE = mae), ...)

\method{modelmetrics}{Surv}(x, y, metrics = c(CIndex = cindex, ROC =
  roc_auc, Brier = brier), ...)
}
\arguments{
\item{x}{observed responses or class containing observed and predicted
responses.}

\item{...}{arguments passed to the \code{\link{metrics}} functions and from
the \code{Resamples} method to the others.}

\item{na.rm}{logical indicating whether to remove observed or predicted
responses that are \code{NA} when calculating model metrics.}

\item{y}{predicted responses.}

\item{metrics}{function, one or more function names, or list of named
functions to include in the calculation of performance metrics.}
}
\description{
Compute measures of model performance.
}
\examples{
res <- resample(Species ~ ., data = iris, model = GBMModel)
(metrics <- modelmetrics(res))
summary(metrics)
plot(metrics)

## Survival response example
library(survival)
library(MASS)

fo <- Surv(time, status != 2) ~ sex + age + year + thickness + ulcer
gbmfit <- fit(fo, data = Melanoma, model = GBMModel)

obs <- response(fo, data = Melanoma)
pred <- predict(gbmfit, newdata = Melanoma, type = "prob")
modelmetrics(obs, pred)

}
\seealso{
\code{\link{metrics}}, \code{\link{resample}},
\code{\link{Resamples}}

\code{\link{predict}}, \code{\link{response}}
}
