% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictors.R
\name{offset}
\alias{offset}
\title{Model Formula Offset}
\usage{
offset(object)
}
\arguments{
\item{object}{offset term to include in a model formula.}
}
\description{
An offset is a term to be added to a linear predictor, such as in a
generalized linear model, with known coefficient 1 rather than an estimated
coefficient.
}
\seealso{
\code{\link[stats:offset]{offset}}
}
