% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot}
\alias{plot}
\alias{plot.MLModelTune}
\alias{plot.Resamples}
\alias{plot.VarImp}
\title{Model Performance Plot}
\usage{
\method{plot}{MLModelTune}(x, metrics = NULL, stat = mean,
  type = c("boxplot", "violin", "errorbar"), ...)

\method{plot}{Resamples}(x, metrics = NULL, stat = mean,
  type = c("boxplot", "violin", "errorbar"), ...)

\method{plot}{VarImp}(x, n = NULL, ...)
}
\arguments{
\item{x}{object to plot.}

\item{metrics}{vector of numeric indexes or character names of the performance
metrics to plot.}

\item{stat}{function to compute a summary statistic on resampled values for
sorting of the plotted models.  The supplied function should contain a
\code{na.rm} argument in its definition.}

\item{type}{type of plot to construct.}

\item{...}{arguments to be passed to other methods.}

\item{n}{number of most important variables to include in the plot
(default: all).}
}
\description{
Plot measures of model performance and predictor variable importance.
}
\seealso{
\code{\link{resample}}, \code{\link{Resamples}}, \code{\link{tune}}

\code{\link{varimp}}
}
