% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLPLSModel.R
\name{PLSModel}
\alias{PLSModel}
\title{Partial Least Squares Model}
\usage{
PLSModel(ncomp = 1, scale = NULL)
}
\arguments{
\item{ncomp}{number of components to include in the model.}

\item{scale}{logical indicating whether to scale the predictors by the
sample standard deviation.}
}
\value{
MLModel class object.
}
\description{
Function to perform partial least squares regression.
}
\details{
\describe{
\item{Response Types:}{\code{factor}, \code{numeric}}
}

Default values for the \code{NULL} arguments and further model details can be
found in the source link below.
}
\seealso{
\code{\link[pls]{mvr}}, \code{\link{fit}}, \code{\link{resample}},
\code{\link{tune}}
}
