% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLPOLRModel.R
\name{POLRModel}
\alias{POLRModel}
\title{Ordered Logistic or Probit Regression Model}
\usage{
POLRModel(method = NULL)
}
\arguments{
\item{method}{logistic or probit or (complementary) log-log or cauchit
(corresponding to a Cauchy latent variable).}
}
\value{
MLModel class object.
}
\description{
Fit a logistic or probit regression model to an ordered factor response.
}
\details{
\describe{
\item{Response Types:}{\code{ordered}}
}

Default values for the \code{NULL} arguments and further model
details can be found in the source link below.
}
\seealso{
\code{\link[MASS]{polr}}, \code{\link{fit}}, \code{\link{resample}},
\code{\link{tune}}
}
