% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print}
\alias{print}
\alias{print.BinomialVariate}
\alias{print.Calibration}
\alias{print.DiscreteVariate}
\alias{print.ListOf}
\alias{print.MLModel}
\alias{print.ModelFrame}
\alias{print.ModeledInput}
\alias{print.Performance}
\alias{print.PerformanceCurve}
\alias{print.RecipeGrid}
\alias{print.Resamples}
\alias{print.SelectedInput}
\alias{print.SurvMatrix}
\alias{print.SurvMeans}
\alias{print.TrainStep}
\alias{print.TunedInput}
\alias{print.VarImp}
\title{Print MachineShop Objects}
\usage{
\method{print}{BinomialVariate}(x, n = MachineShop::settings("print_max"), ...)

\method{print}{Calibration}(x, n = MachineShop::settings("print_max"), ...)

\method{print}{DiscreteVariate}(x, n = MachineShop::settings("print_max"), ...)

\method{print}{ListOf}(x, n = MachineShop::settings("print_max"), ...)

\method{print}{MLModel}(x, n = MachineShop::settings("print_max"), ...)

\method{print}{ModelFrame}(x, n = MachineShop::settings("print_max"), ...)

\method{print}{ModeledInput}(x, n = MachineShop::settings("print_max"), ...)

\method{print}{Performance}(x, n = MachineShop::settings("print_max"), ...)

\method{print}{PerformanceCurve}(x, n = MachineShop::settings("print_max"), ...)

\method{print}{RecipeGrid}(x, n = MachineShop::settings("print_max"), ...)

\method{print}{Resamples}(x, n = MachineShop::settings("print_max"), ...)

\method{print}{SelectedInput}(x, n = MachineShop::settings("print_max"), ...)

\method{print}{SurvMatrix}(x, n = MachineShop::settings("print_max"), ...)

\method{print}{SurvMeans}(x, n = MachineShop::settings("print_max"), ...)

\method{print}{TrainStep}(x, n = MachineShop::settings("print_max"), ...)

\method{print}{TunedInput}(x, n = MachineShop::settings("print_max"), ...)

\method{print}{VarImp}(x, n = MachineShop::settings("print_max"), ...)
}
\arguments{
\item{x}{object to print.}

\item{n}{integer number of models or data frame rows to show.}

\item{...}{arguments passed to other methods.}
}
\description{
Print methods for objects defined in the \pkg{MachineShop} package.
}
