% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.R
\name{print}
\alias{print}
\alias{print.MLModelTune}
\alias{print.Performance}
\alias{print.RecipeGrid}
\alias{print.Resamples}
\alias{print.SurvMatrix}
\alias{print.TunedRecipe}
\alias{print.VarImp}
\title{Print MachineShop Objects}
\usage{
\method{print}{MLModelTune}(x, n = MachineShop::settings("max.print"),
  ...)

\method{print}{Performance}(x, n = MachineShop::settings("max.print"),
  ...)

\method{print}{RecipeGrid}(x, n = MachineShop::settings("max.print"),
  ...)

\method{print}{Resamples}(x, n = MachineShop::settings("max.print"), ...)

\method{print}{SurvMatrix}(x, n = MachineShop::settings("max.print"),
  ...)

\method{print}{TunedRecipe}(x, n = MachineShop::settings("max.print"),
  ...)

\method{print}{VarImp}(x, n = MachineShop::settings("max.print"), ...)
}
\arguments{
\item{x}{object to print.}

\item{n}{integer number of models or data frame rows to show.}

\item{...}{arguments passed to other methods.}
}
\description{
Print methods for objects defined in the \pkg{MachineShop} package.
}
