% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print}
\alias{print}
\alias{print.BinomialVariate}
\alias{print.Calibration}
\alias{print.DiscreteVariate}
\alias{print.ListOf}
\alias{print.MLControl}
\alias{print.MLMetric}
\alias{print.MLModel}
\alias{print.MLModelFunction}
\alias{print.ModelFrame}
\alias{print.ModelRecipe}
\alias{print.ModelSpecification}
\alias{print.Performance}
\alias{print.PerformanceCurve}
\alias{print.RecipeGrid}
\alias{print.Resample}
\alias{print.SurvMatrix}
\alias{print.SurvTimes}
\alias{print.TrainingStep}
\alias{print.VariableImportance}
\title{Print MachineShop Objects}
\usage{
\method{print}{BinomialVariate}(x, n = MachineShop::settings("print_max"), ...)

\method{print}{Calibration}(x, n = MachineShop::settings("print_max"), ...)

\method{print}{DiscreteVariate}(x, n = MachineShop::settings("print_max"), ...)

\method{print}{ListOf}(x, n = MachineShop::settings("print_max"), ...)

\method{print}{MLControl}(x, n = MachineShop::settings("print_max"), ...)

\method{print}{MLMetric}(x, ...)

\method{print}{MLModel}(x, n = MachineShop::settings("print_max"), id = FALSE, ...)

\method{print}{MLModelFunction}(x, ...)

\method{print}{ModelFrame}(x, n = MachineShop::settings("print_max"), id = FALSE, ...)

\method{print}{ModelRecipe}(x, n = MachineShop::settings("print_max"), id = FALSE, ...)

\method{print}{ModelSpecification}(x, n = MachineShop::settings("print_max"), ...)

\method{print}{Performance}(x, n = MachineShop::settings("print_max"), ...)

\method{print}{PerformanceCurve}(x, n = MachineShop::settings("print_max"), ...)

\method{print}{RecipeGrid}(x, n = MachineShop::settings("print_max"), ...)

\method{print}{Resample}(x, n = MachineShop::settings("print_max"), ...)

\method{print}{SurvMatrix}(x, n = MachineShop::settings("print_max"), ...)

\method{print}{SurvTimes}(x, n = MachineShop::settings("print_max"), ...)

\method{print}{TrainingStep}(x, n = MachineShop::settings("print_max"), ...)

\method{print}{VariableImportance}(x, n = MachineShop::settings("print_max"), ...)
}
\arguments{
\item{x}{object to print.}

\item{n}{integer number of models or data frame rows to show.}

\item{...}{arguments passed to other methods, including the one described
below.
\describe{
  \item{\code{level} = 0}{current nesting level of the corresponding
  object in recursive calls to \code{print}.  The amount of information
  displayed decreases and increases with positive and negative levels,
  respectively.}
}}

\item{id}{logical indicating whether to show object identifiers.}
}
\description{
Print methods for objects defined in the \pkg{MachineShop} package.
}
