% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{Resamples}
\alias{Resamples}
\title{Resamples Class Contructor}
\usage{
Resamples(..., control = NULL, strata = character())
}
\arguments{
\item{...}{named or unnamed resample output from one or more models.}

\item{control}{\code{MLControl} object used to generate the resample output.}

\item{strata}{character string indicating the strata variable, if any.}
}
\value{
\code{Resamples} class object.
}
\description{
Create an object of resampled performance metrics from one or more models.
}
\details{
Argument \code{control} need only be specified if the supplied
output is not a \code{Resamples} object.  Output being combined from more
than one model must have been generated with the same resampling control
object.
}
\examples{
## Factor response example

fo <- Species ~ .
control <- CVControl()

gbmres1 <- resample(fo, iris, GBMModel(n.trees = 25), control)
gbmres2 <- resample(fo, iris, GBMModel(n.trees = 50), control)
gbmres3 <- resample(fo, iris, GBMModel(n.trees = 100), control)

res <- Resamples(GBM1 = gbmres1, GBM2 = gbmres2, GBM3 = gbmres3)
summary(res)
plot(res)

}
\seealso{
\code{\link{resample}}, \code{\link{plot}}, \code{\link{summary}}
}
