% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response.R
\name{response}
\alias{response}
\alias{response.formula}
\alias{response.MLModelFit}
\alias{response.ModelFrame}
\alias{response.recipe}
\title{Extract Response Variable}
\usage{
response(object, ...)

\method{response}{formula}(object, data = NULL, ...)

\method{response}{MLModelFit}(object, data = NULL, ...)

\method{response}{ModelFrame}(object, data = NULL, ...)

\method{response}{recipe}(object, data = NULL, ...)
}
\arguments{
\item{object}{object containing response variable values.}

\item{...}{arguments passed to other methods.}

\item{data}{\code{data.frame} containing values at which to evaluate the
response variable.}
}
\description{
Extract the response variable from an object.
}
\examples{
## Survival response example
library(survival)
library(MASS)

fo <- Surv(time, status != 2) ~ sex + age + year + thickness + ulcer
response(fo, data = Melanoma)

}
\seealso{
\code{\link[recipes]{recipe}}
}
