% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response.R
\name{response}
\alias{response}
\alias{response.formula}
\title{Extract Response Variable}
\usage{
response(object, ...)

\method{response}{formula}(object, data, ...)
}
\arguments{
\item{object}{object containing a response.}

\item{...}{arguments passed to other methods.}

\item{data}{data frame containing the values of a response variable defined in a formula.}
}
\description{
Extract the response variable from an object.
}
\examples{
## Survival response example
library(survival)

fo <- Surv(time, status) ~ age + sex + ph.ecog + ph.karno + pat.karno +
                           meal.cal + wt.loss
response(fo, data = lung)

}
