% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response.R
\name{response}
\alias{response}
\alias{response.formula}
\alias{response.MLModelFit}
\alias{response.ModelFrame}
\alias{response.recipe}
\title{Extract Response Variable}
\usage{
response(object, ...)

\method{response}{formula}(object, newdata = NULL, ...)

\method{response}{MLModelFit}(object, newdata = NULL, ...)

\method{response}{ModelFrame}(object, newdata = NULL, ...)

\method{response}{recipe}(object, newdata = NULL, ...)
}
\arguments{
\item{object}{object containing the response variable definition.}

\item{...}{arguments passed to other methods.}

\item{newdata}{data frame from which to extract the response
variable values if given; otherwise, \code{object} is used.}
}
\description{
Extract the response variable from an object.
}
\examples{
## Survival response example
library(survival)
library(MASS)

fo <- Surv(time, status != 2) ~ sex + age + year + thickness + ulcer
response(fo, Melanoma)

}
\seealso{
\code{\link[recipes]{recipe}}
}
