% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ML_SelectedModel.R
\name{SelectedModel}
\alias{SelectedModel}
\title{Selected Model}
\usage{
SelectedModel(..., control = MachineShop::settings("control"),
  metrics = NULL, stat = MachineShop::settings("stat.ModelTune"),
  cutoff = NULL)
}
\arguments{
\item{...}{\link[=models]{model} functions, function names, calls, or vector
of these to serve as the candidate set from which to select.}

\item{control}{\link[=controls]{control} function, function name, or call
defining the resampling method to be employed.}

\item{metrics}{\link[=metrics]{metric} function, function name, or vector of
these with which to calculate performance.  If not specified, default metrics
defined in the \link{performance} functions are used.  Model selection is
based on the first calculated metric.}

\item{stat}{function or character string naming a function to compute a
summary statistic on resampled metric values for model selection.}

\item{cutoff}{argument passed to the \code{metrics} functions.}
}
\value{
\code{SelectedModel} class object that inherits from \code{MLModel}.
}
\description{
Model selected from a candidate set, as produced by the \code{\link{tune}}
function.
}
\details{
\describe{
\item{Response Types:}{\code{factor}, \code{numeric}, \code{ordered},
\code{Surv}
}
}
}
\examples{
fit(sale_amount ~ ., data = ICHomes,
    model = SelectedModel(GBMModel, GLMNetModel, SVMRadialModel))

}
\seealso{
\code{\link{tune}}, \code{\link{fit}}, \code{\link{resample}}
}
