% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLControl.R
\name{set_predict}
\alias{set_predict}
\title{Resampling Prediction Control}
\usage{
set_predict(x, times = NULL, distr = NULL, method = NULL)
}
\arguments{
\item{x}{\link[=controls]{control} object.}

\item{times, distr, method}{arguments passed to \code{\link{predict}}.}
}
\value{
Argument \code{x} updated with the supplied parameters.
}
\description{
Set parameters that control prediction during resample estimation of model
performance.
}
\examples{
CVControl() \%>\% set_predict(times = 1:3)

}
\seealso{
\code{\link{set_strata}},
\code{\link{resample}}, \code{\link{SelectedInput}},
\code{\link{SelectedModel}}, \code{\link{TunedInput}},
\code{\link{TunedModel}}
}
