% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrainingParams.R
\name{set_predict}
\alias{set_predict}
\title{Resampling Prediction Control}
\usage{
set_predict(
  object,
  times = numeric(),
  distr = character(),
  method = character(),
  ...
)
}
\arguments{
\item{object}{\link[=controls]{control} object.}

\item{times, distr, method}{arguments passed to \code{\link{predict}}.}

\item{...}{arguments passed to other methods.}
}
\value{
Argument \code{object} updated with the supplied parameters.
}
\description{
Set parameters that control prediction during resample estimation of model
performance.
}
\examples{
CVControl() \%>\% set_predict(times = 1:3)

}
\seealso{
\code{\link{resample}}, \code{\link{set_monitor}},
\code{\link{set_optim}}, \code{\link{set_strata}}
}
