% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response.R
\name{SurvMatrix}
\alias{SurvMatrix}
\alias{SurvEvents}
\alias{SurvProbs}
\title{SurvMatrix Class Constructors}
\usage{
SurvEvents(object = numeric(), times = NULL)

SurvProbs(object = numeric(), times = NULL)
}
\arguments{
\item{object}{matrix, or object that can be coerced to one, with survival
events or probabilities at points in time in the columns and cases in the
rows.}

\item{times}{numeric vector of survival times for the columns.}
}
\value{
Object that is of the same class as the constructor name and inherits
from \code{SurvMatrix}.  Examples of these are predicted survival events and
probabilities returned by the \link{predict} function.
}
\description{
Create a matrix of survival events or probabilites.
}
\seealso{
\code{\link{performance}}, \code{\link{metrics}}
}
