% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLSVMModel.R
\name{SVMModel}
\alias{SVMModel}
\title{Support Vector Machine Model}
\usage{
SVMModel(scaled = NULL, type = NULL, kernel = NULL, kpar = NULL,
  C = NULL, nu = NULL, epsilon = NULL, cache = NULL, tol = NULL,
  shrinking = NULL)
}
\arguments{
\item{scaled}{logical vector indicating the variables to be scaled.}

\item{type}{type of support vector machine.}

\item{kernel}{kernel function used in training and predicting.}

\item{kpar}{list of hyper-parameters (kernel parameters).}

\item{C}{cost of constraints violation defined as the regularization term in
the Lagrange formulation.}

\item{nu}{parameter needed for nu-svc, one-svc, and nu-svr.}

\item{epsilon}{parameter in the insensitive-loss function used for eps-svr,
nu-svr and eps-bsvm}

\item{cache}{cache memory in MB.}

\item{tol}{tolerance of termination criterion.}

\item{shrinking}{whether to use the shrinking-heuristics.}
}
\description{
Fits the well known C-svc, nu-svc, (classification) one-class-svc (novelty)
eps-svr, nu-svr (regression) formulations along with native multi-class
classification formulations and the bound-constraint SVM formulations.
}
\details{
\describe{
\item{Response Types:}{\code{factor}, \code{numeric}}
}

Default values for the \code{NULL} arguments and further model
details can be found in the source link below.
}
\seealso{
\code{\link[kernlab]{ksvm}}
}
