% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ML_SVMModel.R
\name{SVMModel}
\alias{SVMModel}
\alias{SVMANOVAModel}
\alias{SVMBesselModel}
\alias{SVMLaplaceModel}
\alias{SVMLinearModel}
\alias{SVMPolyModel}
\alias{SVMRadialModel}
\alias{SVMSplineModel}
\alias{SVMTanhModel}
\title{Support Vector Machine Models}
\usage{
SVMModel(
  scaled = TRUE,
  type = character(),
  kernel = c("rbfdot", "polydot", "vanilladot", "tanhdot", "laplacedot", "besseldot",
    "anovadot", "splinedot"),
  kpar = "automatic",
  C = 1,
  nu = 0.2,
  epsilon = 0.1,
  prob.model = FALSE,
  cache = 40,
  tol = 0.001,
  shrinking = TRUE
)

SVMANOVAModel(sigma = 1, degree = 1, ...)

SVMBesselModel(sigma = 1, order = 1, degree = 1, ...)

SVMLaplaceModel(sigma = numeric(), ...)

SVMLinearModel(...)

SVMPolyModel(degree = 1, scale = 1, offset = 1, ...)

SVMRadialModel(sigma = numeric(), ...)

SVMSplineModel(...)

SVMTanhModel(scale = 1, offset = 1, ...)
}
\arguments{
\item{scaled}{logical vector indicating the variables to be scaled.}

\item{type}{type of support vector machine.}

\item{kernel}{kernel function used in training and predicting.}

\item{kpar}{list of hyper-parameters (kernel parameters).}

\item{C}{cost of constraints violation defined as the regularization term in
the Lagrange formulation.}

\item{nu}{parameter needed for nu-svc, one-svc, and nu-svr.}

\item{epsilon}{parameter in the insensitive-loss function used for eps-svr,
nu-svr and eps-bsvm.}

\item{prob.model}{logical indicating whether to calculate the scaling
parameter of the Laplacian distribution fitted on the residuals of numeric
response variables.  Ignored in the case of a factor response variable.}

\item{cache}{cache memory in MB.}

\item{tol}{tolerance of termination criterion.}

\item{shrinking}{whether to use the shrinking-heuristics.}

\item{sigma}{inverse kernel width used by the ANOVA, Bessel, and Laplacian
kernels.}

\item{degree}{degree of the ANOVA, Bessel, and polynomial kernel functions.}

\item{...}{arguments passed to \code{SVMModel} from the other constructors.}

\item{order}{order of the Bessel function to be used as a kernel.}

\item{scale}{scaling parameter of the polynomial and hyperbolic tangent
kernels as a convenient way of normalizing patterns without the need to
modify the data itself.}

\item{offset}{offset used in polynomial and hyperbolic tangent kernels.}
}
\value{
\code{MLModel} class object.
}
\description{
Fits the well known C-svc, nu-svc, (classification) one-class-svc (novelty)
eps-svr, nu-svr (regression) formulations along with native multi-class
classification formulations and the bound-constraint SVM formulations.
}
\details{
\describe{
  \item{Response types:}{\code{factor}, \code{numeric}}
  \item{\link[=TunedModel]{Automatic tuning} of grid parameters:}{
    \itemize{
      \item SVMModel: \code{NULL}
      \item SVMANOVAModel: \code{C}, \code{degree}
      \item SVMBesselModel: \code{C}, \code{order}, \code{degree}
      \item SVMLaplaceModel: \code{C}, \code{sigma}
      \item SVMLinearModel: \code{C}
      \item SVMPolyModel: \code{C}, \code{degree}, \code{scale}
      \item SVMRadialModel: \code{C}, \code{sigma}
    }
  }
}

The kernel-specific constructor functions \code{SVMANOVAModel},
\code{SVMBesselModel}, \code{SVMLaplaceModel}, \code{SVMLinearModel},
\code{SVMPolyModel}, \code{SVMRadialModel}, \code{SVMSplineModel}, and
\code{SVMTanhModel} are special cases of \code{SVMModel} which automatically
set its \code{kernel} and \code{kpar} arguments.  These are called directly
in typical usage unless \code{SVMModel} is needed to specify a more general
model.

Default argument values and further model details can be found in the source
See Also link below.
}
\examples{
fit(sale_amount ~ ., data = ICHomes, model = SVMRadialModel)

}
\seealso{
\code{\link[kernlab]{ksvm}}, \code{\link{fit}},
\code{\link{resample}}
}
