% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune.R
\name{tune}
\alias{tune}
\alias{tune.formula}
\alias{tune.matrix}
\alias{tune.ModelFrame}
\alias{tune.recipe}
\alias{tune.MLModel}
\alias{tune.MLModelFunction}
\title{Model Tuning and Selection}
\usage{
tune(x, ...)

\method{tune}{formula}(x, data, model, ...)

\method{tune}{matrix}(x, y, model, ...)

\method{tune}{ModelFrame}(x, model, ...)

\method{tune}{recipe}(x, model, ...)

\method{tune}{MLModel}(x, ...)

\method{tune}{MLModelFunction}(x, ...)
}
\arguments{
\item{x}{defines a relationship between model predictor and response
variables.  May be a \code{\link{formula}}, design \code{\link{matrix}} of
predictors, \code{\link{ModelFrame}}, untrained
\code{\link[recipes]{recipe}}, \code{\link{SelectedRecipe}}, or
\code{\link{TunedRecipe}} object. Alternatively, a \link[=models]{model}
function or call may be given first followed by objects defining the
predictor and response relationship and the other tuning argument values.}

\item{...}{arguments passed to other methods.}

\item{data}{\link[=data.frame]{data frame} containing observed predictors and
outcomes.}

\item{model}{\link[=models]{model} function, function name, or call.  Supply
a \code{\link{SelectedModel}} object for model selection or a
\code{\link{TunedModel}} object for model tuning.}

\item{y}{response variable.}
}
\value{
\code{MLModel} class object containing the tuning results.
}
\description{
Predictive peformance-based tuning of a model over a grid of parameters
values or selection from a set of candidate models.
}
\details{
The \code{tune} function is intended for internal use by the package and will
be deprecated in the future.  To perform model tuning or selection, users
should call the \code{\link{fit}} function with a \code{\link{TunedModel}} or
\code{\link{SelectedModel}} instead.
}
\seealso{
\code{\link{fit}}, \code{\link{SelectedModel}},
\code{\link{TunedModel}}
}
