% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune.R
\docType{methods}
\name{tune}
\alias{tune}
\alias{tune,function,data.frame-method}
\alias{tune,function,formula-method}
\alias{tune,function,recipe-method}
\title{Model Tuning}
\usage{
tune(object, x, ...)

\S4method{tune}{`function`,data.frame}(object, x, control,
  grid = data.frame(), metric = 1, stat = mean, maximize = TRUE)

\S4method{tune}{`function`,formula}(object, x, data, control,
  grid = data.frame(), metric = 1, stat = mean, maximize = TRUE)

\S4method{tune}{`function`,recipe}(object, x, control,
  grid = data.frame(), metric = 1, stat = mean, maximize = TRUE)
}
\arguments{
\item{object}{constructor function for the model.}

\item{x}{defined relationship between model predictors and an outcome.  May
be a model.frame (data.frame) containing a formula, data, and optionally case
weights; a formula; or a recipe.}

\item{...}{further arguments passed to other methods.}

\item{control}{\code{\linkS4class{MLControl}} object defining and controlling
the resampling method to be employed.}

\item{grid}{data frame containing parameter values over which to evaluate the
\code{object} model function.}

\item{metric}{numeric index or character name of the performance metric to
use in selecting the best model.}

\item{stat}{function to compute a summary statistic on resampled values of
the metric for model selection.  The supplied function should contain a
\code{na.rm} argument in its definition.}

\item{maximize}{logical indicating whether to select the model having the
maximum or minimum value of the performance metric.}

\item{data}{data frame containing observed predictors and outcomes.}
}
\value{
MLModelTune class object that inherits from MLModel.
}
\description{
Evaluate a model over a grid of tuning parameters and select the best one
according to resample estimation of predictive performance.
}
\seealso{
\code{\link[stats]{model.frame}}, \code{\link[recipes]{recipe}},
\code{\link{BootControl}}, \code{\link{CVControl}}, \code{\link{OOBControl}},
\code{\link{fit}}, \code{\link{resample}}
}
