% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varimp.R
\name{varimp}
\alias{varimp}
\title{Variable Importance}
\usage{
varimp(object, scale = TRUE, ...)
}
\arguments{
\item{object}{MLModelFit object from a model fit.}

\item{scale}{logical indicating whether importance measures should be scaled
to range from 0 to 100.}

\item{...}{arguments passed to model-specific variable importance functions.}
}
\description{
Calculate measures of the relative importance of predictors in a model.
}
\seealso{
\code{\link{fit}}, \code{\link[MLModels:plot.VarImp]{plot}}
}
